renamed tests/web/common.py to 'webcommon.py' and refactored accordingly, the module gets loaded and conflicts with tests/app/common.py
